ImagePresentation: move ViewRenderInputs to header
